更改id类型uint->int64

+++ controller/admin_account/admin_change_password_handle.go
@@ -11,7 +11,7 @@ import (
 // AdminChangePasswordHandle 将用户名对应的管理员旧密码更改为新密码
 // 1. 比对输入旧密码与数据库中密码是否相同 2. 将数据库中密码更新为新密码
 func AdminChangePasswordHandle(context *gin.Context) {
-	UserID, ok := context.MustGet("user_id").(uint)
+	UserID, ok := context.MustGet("user_id").(int64)
 	if !ok {
 		context.JSON(http.StatusInternalServerError, common.NewCommonResponse(1, "解析用户信息错误"))
 		return
